Clean up .js extensions, path aliases
