+++ app/frontend/components/SyncIndicator.jsx
