+++ app/frontend/stylesheets/directive.scss
@@ -2062,7 +2062,7 @@
       -webkit-transition: opacity 0.25s ease-in-out;
       -ms-transition: opacity 0.25s ease-in-out;
       transition: opacity 0.25s ease-in-out;
-      background-image: url("@images/overlay.png");
+      background-image: url("~images/overlay.png");
       content: '';
       display: block;
       height: 100%;
@@ -2473,7 +2473,7 @@
 
   #header {
     background-attachment: scroll,              scroll,          fixed;
-    background-image: url("@images/top-3200.svg"), url("@images/overlay.png"), url("@images/annie-spratt-197448-web.jpg");
+    background-image: url("~images/top-3200.svg"), url("~images/overlay.png"), url("~images/annie-spratt-197448-web.jpg");
     background-position: bottom center,          top left,        center center;
     background-repeat: repeat-x,            repeat,          no-repeat;
     background-size: 3200px 460px,          auto,          cover;
@@ -2580,7 +2580,7 @@
   #footer {
     background-attachment: scroll;
     background-color: #4eb980;
-    background-image: url("@images/bottom-3200.svg");
+    background-image: url("~images/bottom-3200.svg");
     background-position: top center;
     background-repeat: repeat-x;
     background-size: 3200px 460px;
@@ -2805,7 +2805,7 @@
 
       #header {
         background-attachment: scroll,              scroll,              scroll;
-        background-image: url("@images/top-1600.svg"), url("@images/overlay.png"), url("@images/annie-spratt-197448-web.jpg");
+        background-image: url("~images/top-1600.svg"), url("~images/overlay.png"), url("~images/annie-spratt-197448-web.jpg");
         background-size: 1600px 230px,          auto,              cover;
         padding: 6em 0 12em 0;
       }
@@ -2829,7 +2829,7 @@
     /* Footer */
 
       #footer {
-        background-image: url("@images/bottom-1600.svg");
+        background-image: url("~images/bottom-1600.svg");
         background-size: 1600px 230px;
         margin-top: -100px;
         padding: calc(230px + 4em) 0 6em 0;
@@ -2967,7 +2967,7 @@
     /* Header */
 
       #header {
-        background-image: url("@images/top-1280.svg"), url("@images/overlay.png"), url("@images/annie-spratt-197448-web.jpg");
+        background-image: url("~images/top-1280.svg"), url("~images/overlay.png"), url("~images/annie-spratt-197448-web.jpg");
         background-size: 1280px 184px,            auto,              cover;
         padding: 3em 0 7em 0;
       }
@@ -2995,7 +2995,7 @@
     /* Footer */
 
       #footer {
-        background-image: url("@images/bottom-1280.svg");
+        background-image: url("~images/bottom-1280.svg");
         background-size: 1280px 184px;
         margin-top: -75px;
         padding: calc(184px + 2em) 0 2em 0;
